Move server tests to functional set of tests
